+++ darkskyapi/src/main/java/org/ups/greensky/api/darksky/result/DailyItem.java
@@ -95,7 +95,7 @@ public class DailyItem {
     private Long uvIndexTime;
     @SerializedName("visibility")
     @Expose
-    private Long visibility;
+    private Double visibility;
     @SerializedName("ozone")
     @Expose
     private Double ozone;
@@ -356,11 +356,11 @@ public class DailyItem {
         this.uvIndexTime = uvIndexTime;
     }
 
-    public Long getVisibility() {
+    public Double getVisibility() {
         return visibility;
     }
 
-    public void setVisibility(Long visibility) {
+    public void setVisibility(Double visibility) {
         this.visibility = visibility;
     }
